+++ src/main/java/com/leiteria/repository/RegraRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.leiteria.model.ERegra;
 import com.leiteria.model.Regra;
 
-public interface RegraRepository extends JpaRepository<Regra, Integer> {
+public interface RegraRepository extends JpaRepository<Regra, Long> {
 	Optional<Regra> findBynomeRegra(ERegra nome);
 
 }
